Refactor upstream clients
