+++ src/BaGet.Core/Upstream/IMirrorService.cs
