+++ src/BaGet.Core/Upstream/PackageDownloadsJsonSource.cs
